moved firebase auth util class
